Update Database.php
Changed charset and collation to check for empty values instead of using the Null Coalescing Operator. The issue would arise when the values for the environment are not present resulting with empty string value for as getenv() will never return NULL value, making it impossible to ever reach the set default.